python-noticeboard/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic==2.0.3
pydantic-settings
mysql-connector-python
passlib[bcrypt]